+++ src/screens/courses/cardTemplates/TextMediaCard/index.tsx
@@ -79,6 +79,6 @@ const TextMediaCard = ({
   );
 };
 
-const mapStateToProps = (state: StateType) => ({ card: Selectors.getCard(state), index: state.activities.cardIndex });
+const mapStateToProps = (state: StateType) => ({ card: Selectors.getCard(state), index: state.cards.cardIndex });
 
 export default connect(mapStateToProps)(TextMediaCard);
