+++ src/screens/courses/cardTemplates/TitleTextMediaCard/index.tsx
@@ -81,6 +81,6 @@ const TitleTextMediaCard = ({
   );
 };
 
-const mapStateToProps = (state: StateType) => ({ card: Selectors.getCard(state), index: state.activities.cardIndex });
+const mapStateToProps = (state: StateType) => ({ card: Selectors.getCard(state), index: state.cards.cardIndex });
 
 export default connect(mapStateToProps)(TitleTextMediaCard);
