+++ src/types/store/ActivityStoreType.ts
@@ -3,7 +3,6 @@ import { ActivityHistoryType } from '../ActivityHistoryType';
 
 // Actions types
 export const SET_ACTIVITY = 'SET_ACTIVITY';
-export const SET_CARD_INDEX = 'SET_CARD_INDEX';
 export const SET_EXIT_CONFIRMATION_MODAL = 'SET_EXIT_CONFIRMATION_MODAL';
 export const ADD_QUESTIONNAIRE_ANSWER = 'ADD_QUESTIONNAIRE_ANSWER';
 export const REMOVE_QUESTIONNAIRE_ANSWER = 'REMOVE_QUESTIONNAIRE_ANSWER';
@@ -15,10 +14,6 @@ export const SET_ACTIVITY_HISTORIES = 'SET_ACTIVITY_HISTORIES';
 export interface SetActivityType {
   type: typeof SET_ACTIVITY,
   payload: ActivityType,
-}
-export interface SetCardIndexType {
-  type: typeof SET_CARD_INDEX,
-  payload: number,
 }
 export interface SetExitConfirmationModalType {
   type: typeof SET_EXIT_CONFIRMATION_MODAL,
@@ -55,7 +50,6 @@ export interface IncGoodAnswersCountType {
 
 export type ActivityActionType =
 SetActivityType |
-SetCardIndexType |
 SetExitConfirmationModalType |
 AddQuestionnaireAnswerType |
 RemoveQuestionnaireAnswerType |
@@ -72,7 +66,6 @@ export interface QuestionnaireAnswerType {
 
 export interface ActivityStateType {
   activity: ActivityType | null,
-  cardIndex: number | null,
   exitConfirmationModal: boolean,
   questionnaireAnswersList: Array<QuestionnaireAnswerType>,
   score: number
